Attempt TraversableGrid problem
Unnecessarily difficult. Leaving there for now as I've spent what must be at least 15 hours on this. Not mad far off, but also been thinking that for a while now. I mean, the generation of the GridDepthTwo and the traversing itself should both be easy. Just, somehow, figuring out the rectangle(s) that a blank square is a part of is not so.